be able to push schema to prod database
turns out planetscale doesn't currently support the "deploy request"
feature that i loved so much from mysql/vitess for postgres. this means
we have to do old-school migrations. no worries - that will be fine. but
that does mean we need the ability to push the schema directly to prod,
which i have now implemented.